Login view
Asks for username and password, check them.
When validation succeeds, the user’s id is stored in a new session. The data is stored in a cookie that is sent to the browser, and the browser then sends it back with subsequent requests. Flask securely signs the data so that it can’t be tampered with.
At the beginning of each request, if a user is logged in their information should be loaded and made available to other views.